move file rotation logic to separate file
