+++ cmd/acra-rotate/acra-rotate.go
@@ -19,19 +19,13 @@ limitations under the License.
 package main
 
 import (
-	"encoding/json"
 	"flag"
-	"fmt"
-	"github.com/cossacklabs/acra/acra-writer"
 	"github.com/cossacklabs/acra/cmd"
-	"github.com/cossacklabs/acra/decryptor/base"
 	"github.com/cossacklabs/acra/keystore"
 	"github.com/cossacklabs/acra/keystore/filesystem"
 	"github.com/cossacklabs/acra/logging"
 	"github.com/cossacklabs/acra/utils"
-	"github.com/cossacklabs/themis/gothemis/keys"
 	log "github.com/sirupsen/logrus"
-	"io/ioutil"
 	"os"
 )
 
@@ -66,78 +60,9 @@ func initKeyStore(dirPath string) (keystore.KeyStore, error) {
 	return keystorage, nil
 }
 
-func loadFileMap(path string) (ZoneIDFileMap, error) {
-	configData, err := ioutil.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-	return ParseConfig(configData)
-}
-
-// ZoneRotateData store new public key and paths of files that was rotated
-type ZoneRotateData struct {
-	NewPublicKey []byte   `json:"new_public_key"`
-	FilePaths    []string `json:"file_paths"`
-}
-
-// ZoneRotateResult store result of rotation
-type ZoneRotateResult map[string]*ZoneRotateData
-
-func rotateFiles(fileMap ZoneIDFileMap, keyStore keystore.KeyStore) (ZoneRotateResult, error) {
-	output := ZoneRotateResult{}
-	for zoneID, paths := range fileMap {
-		logger := log.WithField("zone_id", zoneID)
-		binZoneID := []byte(zoneID)
-		privateKey, err := keyStore.GetZonePrivateKey(binZoneID)
-		if err != nil {
-			logger.WithError(err).Errorln("Can't load private key of zone")
-			return nil, err
-		}
-		newPublicKey, err := keyStore.RotateZoneKey(binZoneID)
-		if err != nil {
-			logger.WithError(err).Errorln("Can't rotate zone key")
-			return nil, err
-		}
-		result := &ZoneRotateData{NewPublicKey: newPublicKey}
-		for _, path := range paths {
-			fileLogger := logger.WithField("filepath", path)
-			result.FilePaths = append(result.FilePaths, path)
-			acraStruct, err := ioutil.ReadFile(path)
-			if err != nil {
-				fileLogger.WithError(err).Errorf("Can't read file %s", path)
-				return nil, err
-			}
-			decrypted, err := base.DecryptAcrastruct(acraStruct, privateKey, binZoneID)
-			if err != nil {
-				fileLogger.WithError(err).Errorln("Can't decrypt AcraStruct")
-				return nil, err
-			}
-			rotated, err := acrawriter.CreateAcrastruct(decrypted, &keys.PublicKey{Value: newPublicKey}, binZoneID)
-			if err != nil {
-				fileLogger.WithError(err).Errorln("Can't re-encrypt AcraStruct with rotated zone key")
-				return nil, err
-			}
-			stat, err := os.Stat(path)
-			if err != nil {
-				fileLogger.WithError(err).Errorln("Can't get stat info about file to retrieve current file permissions")
-				return nil, err
-			}
-			if err := ioutil.WriteFile(path, rotated, stat.Mode()); err != nil {
-				fileLogger.WithError(err).Errorln("Can't write rotated AcraStruct with zone")
-				return nil, err
-			}
-			fileLogger.Infof("Finish rotate file")
-		}
-		output[zoneID] = result
-		logger.Infoln("Finish rotate zone")
-	}
-	return output, nil
-}
-
 func main() {
 	keysDir := flag.String("keys_dir", keystore.DefaultKeyDirShort, "Folder from which the keys will be loaded")
-	fileMapConfig := flag.String("file_map_config", "", "Path to file with map of <ZoneId>: <FilePath>")
-	//clientID := flag.String("client_id", "", "Client ID should be name of file with private key")
+	fileMapConfig := flag.String("file_map_config", "", "Path to file with map of <ZoneId>: <FilePaths> in json format {\"zone_id1\": [\"filepath1\", \"filepath2\"], \"zone_id2\": [\"filepath1\", \"filepath2\"]}")
 
 	logging.SetLogLevel(logging.LOG_VERBOSE)
 
@@ -147,25 +72,11 @@ func main() {
 		os.Exit(1)
 	}
 
-	//cmd.ValidateClientID(*clientID)
 	keystorage, err := initKeyStore(*keysDir)
 	if err != nil {
 		os.Exit(1)
 	}
-	fileMap, err := loadFileMap(*fileMapConfig)
-	if err != nil {
-		log.WithError(err).Errorln("Can't load config with map <ZoneId>: <FilePath>")
-		os.Exit(1)
+	if *fileMapConfig != "" {
+		runFileRotation(*fileMapConfig, keystorage)
 	}
-	result, err := rotateFiles(fileMap, keystorage)
-	if err != nil {
-		log.WithError(err).Errorln("Can't rotate files")
-		os.Exit(1)
-	}
-	jsonOutput, err := json.Marshal(result)
-	if err != nil {
-		log.WithError(err).Errorln("Can't encode result to json format")
-		os.Exit(1)
-	}
-	fmt.Println(string(jsonOutput))
 }
